+++ src/dfx/src/commands/start.rs
@@ -1,7 +1,6 @@
 use crate::actors::pocketic_proxy::{signals::PortReadySubscribe, PocketIcProxyConfig};
 use crate::actors::{
-    start_btc_adapter_actor, start_canister_http_adapter_actor, start_pocketic_actor,
-    start_pocketic_proxy_actor, start_post_start_actor, start_replica_actor,
+    start_pocketic_actor, start_pocketic_proxy_actor, start_post_start_actor,
     start_shutdown_controller,
 };
 use crate::config::dfx_version_str;
@@ -13,11 +12,10 @@ use crate::lib::network::id::write_network_id;
 use crate::lib::replica::status::ping_and_wait;
 use crate::util::get_reusable_socket_addr;
 use actix::Recipient;
-use anyhow::{anyhow, bail, Context, Error};
+use anyhow::{anyhow, bail, ensure, Context, Error};
 use clap::{ArgAction, Parser};
 use dfx_core::{
     config::model::{
-        bitcoin_adapter, canister_http_adapter,
         local_server_descriptor::{LocalNetworkScopeDescriptor, LocalServerDescriptor},
         network_descriptor::NetworkDescriptor,
         replica_config::{CachedConfig, ReplicaConfig},
@@ -28,13 +26,12 @@ use dfx_core::{
     network::provider::{create_network_descriptor, LocalBindDetermination},
 };
 use fn_error_context::context;
-use os_str_bytes::{OsStrBytes, OsStringBytes};
 use slog::{info, warn, Logger};
 use std::io::Read;
 use std::net::SocketAddr;
 use std::path::{Path, PathBuf};
 use std::process::Command;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::{Duration, SystemTime};
 use sysinfo::{Pid, System, SystemExt};
 use tokio::runtime::Runtime;
 
@@ -65,7 +62,7 @@ pub struct StartOpts {
     #[arg(long)]
     enable_bitcoin: bool,
 
-    /// enable canister http requests (on by default for --pocketic)
+    /// enable canister http requests (on by default)
     #[arg(long)]
     enable_canister_http: bool,
 
@@ -87,8 +84,7 @@ pub struct StartOpts {
     #[allow(unused)]
     pocketic: bool,
 
-    /// Runs the replica instead of pocketic.
-    #[clap(long, conflicts_with = "pocketic")]
+    #[clap(long, hide = true)]
     replica: bool,
 }
 
@@ -160,6 +156,8 @@ pub fn exec(
         replica,
     }: StartOpts,
 ) -> DfxResult {
+    ensure!(!replica, "The 'native' replica (--replica) is no longer supported. See the 0.27.0 migration guide for more information.
+https://github.com/dfinity/sdk/blob/0.27.0/docs/migration/dfx-0.27.0-migration-guide.md");
     if !background {
         info!(
             env.get_logger(),
@@ -192,7 +190,6 @@ pub fn exec(
         enable_canister_http,
         domain,
         artificial_delay,
-        replica,
     )?;
 
     let local_server_descriptor = network_descriptor.local_server_descriptor()?;
@@ -229,20 +226,7 @@ pub fn exec(
 
     clean_older_state_dirs(local_server_descriptor)?;
     let state_root = local_server_descriptor.state_dir();
-
-    let btc_adapter_socket_holder_path = local_server_descriptor.btc_adapter_socket_holder_path();
-    let canister_http_adapter_socket_holder_path =
-        local_server_descriptor.canister_http_adapter_socket_holder_path();
-
     let pid_file_path = empty_writable_path(pid_file_path)?;
-    let btc_adapter_pid_file_path =
-        empty_writable_path(local_server_descriptor.btc_adapter_pid_path())?;
-    let btc_adapter_config_path =
-        empty_writable_path(local_server_descriptor.btc_adapter_config_path())?;
-    let canister_http_adapter_pid_file_path =
-        empty_writable_path(local_server_descriptor.canister_http_adapter_pid_path())?;
-    let canister_http_adapter_config_path =
-        empty_writable_path(local_server_descriptor.canister_http_adapter_config_path())?;
     let pocketic_proxy_pid_file_path =
         empty_writable_path(local_server_descriptor.pocketic_proxy_pid_path())?;
     let pocketic_proxy_port_file_path =
@@ -251,12 +235,6 @@ pub fn exec(
 
     let previous_config_path = local_server_descriptor.effective_config_path();
 
-    // dfx info replica-port will read these port files to find out which port to use,
-    // so we need to make sure only one has a valid port in it.
-    let replica_config_dir = local_server_descriptor.replica_configuration_dir();
-    fs::create_dir_all(&replica_config_dir)?;
-
-    let replica_port_path = empty_writable_path(local_server_descriptor.replica_port_path())?;
     let pocketic_port_path = empty_writable_path(local_server_descriptor.pocketic_port_path())?;
 
     if background {
@@ -278,23 +256,6 @@ pub fn exec(
     write_pid(&pid_file_path);
     fs::write(&webserver_port_path, address_and_port.port().to_string())?;
 
-    let btc_adapter_config = configure_btc_adapter_if_enabled(
-        local_server_descriptor,
-        &btc_adapter_config_path,
-        &btc_adapter_socket_holder_path,
-    )?;
-    let btc_adapter_socket_path = btc_adapter_config
-        .as_ref()
-        .and_then(|cfg| cfg.get_socket_path());
-
-    let canister_http_adapter_config = configure_canister_http_adapter_if_enabled(
-        local_server_descriptor,
-        &canister_http_adapter_config_path,
-        &canister_http_adapter_socket_holder_path,
-    )?;
-    let canister_http_socket_path = canister_http_adapter_config
-        .as_ref()
-        .and_then(|cfg| cfg.get_socket_path());
     let subnet_type = local_server_descriptor
         .replica
         .subnet_type
@@ -311,24 +272,13 @@ pub fn exec(
         .map(|v| v.into_vec());
 
     let replica_config = {
-        let replica_config =
+        let mut replica_config =
             ReplicaConfig::new(&state_root, subnet_type, log_level, artificial_delay);
-        let mut replica_config = if let Some(port) = local_server_descriptor.replica.port {
-            replica_config.with_port(port)
-        } else {
-            replica_config.with_random_port(&replica_port_path)
-        };
-        if let Some(btc_adapter_config) = btc_adapter_config.as_ref() {
+        if local_server_descriptor.bitcoin.enabled {
             replica_config = replica_config.with_btc_adapter_enabled();
-            if let Some(btc_adapter_socket) = btc_adapter_config.get_socket_path() {
-                replica_config = replica_config.with_btc_adapter_socket(btc_adapter_socket);
-            }
         }
-        if let Some(canister_http_adapter_config) = canister_http_adapter_config.as_ref() {
+        if local_server_descriptor.canister_http.enabled {
             replica_config = replica_config.with_canister_http_adapter_enabled();
-            if let Some(socket_path) = canister_http_adapter_config.get_socket_path() {
-                replica_config = replica_config.with_canister_http_adapter_socket(socket_path);
-            }
         }
         replica_config
     };
@@ -364,40 +314,7 @@ pub fn exec(
     let _proxy = system.block_on(async move {
         let shutdown_controller = start_shutdown_controller(env)?;
 
-        let port_ready_subscribe: Recipient<PortReadySubscribe> = if replica {
-            let btc_adapter_ready_subscribe = btc_adapter_config
-                .map(|btc_adapter_config| {
-                    start_btc_adapter_actor(
-                        env,
-                        btc_adapter_config_path,
-                        btc_adapter_config.get_socket_path(),
-                        shutdown_controller.clone(),
-                        btc_adapter_pid_file_path,
-                    )
-                })
-                .transpose()?;
-            let canister_http_adapter_ready_subscribe = canister_http_adapter_config
-                .map(|canister_http_adapter_config| {
-                    start_canister_http_adapter_actor(
-                        env,
-                        canister_http_adapter_config_path,
-                        canister_http_adapter_config.get_socket_path(),
-                        shutdown_controller.clone(),
-                        canister_http_adapter_pid_file_path,
-                    )
-                })
-                .transpose()?;
-
-            let replica = start_replica_actor(
-                env,
-                replica_config,
-                local_server_descriptor,
-                shutdown_controller.clone(),
-                btc_adapter_ready_subscribe,
-                canister_http_adapter_ready_subscribe,
-            )?;
-            replica.recipient()
-        } else {
+        let port_ready_subscribe: Recipient<PortReadySubscribe> = {
             let server = start_pocketic_actor(
                 env,
                 replica_config,
@@ -428,14 +345,6 @@ pub fn exec(
         Ok::<_, Error>(post_start)
     })?;
     system.run()?;
-
-    if let Some(btc_adapter_socket_path) = btc_adapter_socket_path {
-        let _ = std::fs::remove_file(btc_adapter_socket_path);
-    }
-    if let Some(canister_http_socket_path) = canister_http_socket_path {
-        let _ = std::fs::remove_file(canister_http_socket_path);
-    }
-
     Ok(())
 }
 
@@ -496,7 +405,6 @@ pub fn apply_command_line_parameters(
     enable_canister_http: bool,
     domain: Vec<String>,
     artificial_delay: u32,
-    replica: bool,
 ) -> DfxResult<NetworkDescriptor> {
     if enable_canister_http {
         warn!(
@@ -527,12 +435,8 @@ pub fn apply_command_line_parameters(
         local_server_descriptor = local_server_descriptor.with_proxy_domains(domain)
     }
 
-    let settings_digest = get_settings_digest(
-        replica_rev(),
-        &local_server_descriptor,
-        artificial_delay,
-        replica,
-    );
+    let settings_digest =
+        get_settings_digest(replica_rev(), &local_server_descriptor, artificial_delay);
 
     local_server_descriptor = local_server_descriptor.with_settings_digest(settings_digest);
 
@@ -636,108 +540,8 @@ fn write_pid(pid_file_path: &Path) {
     }
 }
 
-#[context("Failed to configure btc adapter.")]
-pub fn configure_btc_adapter_if_enabled(
-    local_server_descriptor: &LocalServerDescriptor,
-    config_path: &Path,
-    uds_holder_path: &Path,
-) -> DfxResult<Option<bitcoin_adapter::Config>> {
-    if !local_server_descriptor.bitcoin.enabled {
-        return Ok(None);
-    };
-
-    let log_level = local_server_descriptor.bitcoin.log_level;
-
-    let nodes = if let Some(ref nodes) = local_server_descriptor.bitcoin.nodes {
-        nodes.clone()
-    } else {
-        bitcoin_adapter::default_nodes()
-    };
-
-    let config = write_btc_adapter_config(uds_holder_path, config_path, nodes, log_level)?;
-    Ok(Some(config))
-}
-
-#[context("Failed to create persistent socket path for {} at {}.", prefix, uds_holder_path.to_string_lossy())]
-fn create_new_persistent_socket_path(uds_holder_path: &Path, prefix: &str) -> DfxResult<PathBuf> {
-    let pid = sysinfo::get_current_pid()
-        .map_err(|s| anyhow!("Unable to obtain pid of current process: {}", s))?;
-    let timestamp_seconds = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
-
-    // Unix domain socket names can only be so long.
-    // An attempt to use a path under .dfx/ resulted in this error:
-    //    path must be shorter than libc::sockaddr_un.sun_path
-    let uds_path = std::env::temp_dir().join(format!("{}.{}.{}", prefix, pid, timestamp_seconds));
-    std::fs::write(uds_holder_path, uds_path.to_raw_bytes()).with_context(|| {
-        format!(
-            "unable to write unix domain socket path to {}",
-            uds_holder_path.to_string_lossy()
-        )
-    })?;
-    Ok(uds_path)
-}
-
-#[context("Failed to get persistent socket path for {} at {}.", prefix, uds_holder_path.to_string_lossy())]
-fn get_persistent_socket_path(uds_holder_path: &Path, prefix: &str) -> DfxResult<PathBuf> {
-    if let Ok(uds_path) = std::fs::read(uds_holder_path) {
-        Ok(PathBuf::assert_from_raw_vec(uds_path))
-    } else {
-        create_new_persistent_socket_path(uds_holder_path, prefix)
-    }
-}
-
-#[context("Failed to write btc adapter config to {}.", config_path.to_string_lossy())]
-fn write_btc_adapter_config(
-    uds_holder_path: &Path,
-    config_path: &Path,
-    nodes: Vec<SocketAddr>,
-    log_level: bitcoin_adapter::BitcoinAdapterLogLevel,
-) -> DfxResult<bitcoin_adapter::Config> {
-    let socket_path = get_persistent_socket_path(uds_holder_path, "ic-btc-adapter-socket")?;
-
-    let adapter_config = bitcoin_adapter::Config::new(nodes, socket_path, log_level);
-
-    let contents = serde_json::to_string_pretty(&adapter_config)
-        .context("Unable to serialize btc adapter configuration to json")?;
-    std::fs::write(config_path, contents).with_context(|| {
-        format!(
-            "Unable to write btc adapter configuration to {}",
-            config_path.to_string_lossy()
-        )
-    })?;
-
-    Ok(adapter_config)
-}
-
 pub fn empty_writable_path(path: PathBuf) -> DfxResult<PathBuf> {
     std::fs::write(&path, "")
         .with_context(|| format!("Unable to write to {}", path.to_string_lossy()))?;
     Ok(path)
-}
-
-#[context("Failed to configure canister http adapter.")]
-pub fn configure_canister_http_adapter_if_enabled(
-    local_server_descriptor: &LocalServerDescriptor,
-    config_path: &Path,
-    uds_holder_path: &Path,
-) -> DfxResult<Option<canister_http_adapter::Config>> {
-    if !local_server_descriptor.canister_http.enabled {
-        return Ok(None);
-    };
-
-    let socket_path =
-        get_persistent_socket_path(uds_holder_path, "ic-https-outcalls-adapter-socket")?;
-
-    let log_level = local_server_descriptor.canister_http.log_level;
-    let adapter_config = canister_http_adapter::Config::new(socket_path, log_level);
-
-    let contents = serde_json::to_string_pretty(&adapter_config)
-        .context("Unable to serialize canister http adapter configuration to json")?;
-    std::fs::write(config_path, contents)
-        .with_context(|| format!("Unable to write {}", config_path.to_string_lossy()))?;
-
-    Ok(Some(adapter_config))
 }