+++ CHANGELOG.md
@@ -6,8 +6,13 @@
 
 * Command `rules` that outputs time-efficient GNU Make rules.
 
+# 0.27.0
+
+### feat!: remove the 'native' replica
+
+The native replica is no longer bundled with dfx; dfx only uses PocketIC for local networks. Accordingly `dfx start --replica` and `dfx info replica-port` now report an error. See the [migration guide](./docs/migration/dfx-0.27.0-migration-guide.md) for more information.
+
 ### chore: removes the outdated `_language-service` command
->>>>>>> upstream/master
 
 ### feat: Support 'follow' mode for 'dfx canister logs'
 Support `follow` mode for `dfx canister logs`
@@ -25,7 +30,20 @@ Improve `dfx canister logs` with several options
 
 ### Motoko
 
-Updated Motoko to [0.14.8](https://github.com/dfinity/motoko/releases/tag/0.14.8)
+Updated Motoko to [0.14.9](https://github.com/dfinity/motoko/releases/tag/0.14.9)
+
+### Bitcoin canister
+
+Upgraded Bitcoin canister to [release/2024-08-30](https://github.com/dfinity/bitcoin-canister/releases/tag/release%2F2024-08-30)
+
+### Replica
+
+Updated replica to elected commit f195ba756bc3bf170a2888699e5e74101fdac6ba.
+This incorporates the following executed proposals:
+
+- [136436](https://dashboard.internetcomputer.org/proposal/136436)
+- [136366](https://dashboard.internetcomputer.org/proposal/136366)
+- [136310](https://dashboard.internetcomputer.org/proposal/136310)
 
 # 0.26.1
 
@@ -52,7 +70,7 @@ Extend `dfx ledger transfer` and `dfx ledger balance` to support [ICRC-1 standar
 
 ### Motoko
 
-Updated Motoko to [0.14.7](https://github.com/dfinity/motoko/releases/tag/0.14.7)
+Updated Motoko to [0.14.8](https://github.com/dfinity/motoko/releases/tag/0.14.8)
 
 ### Replica